requirements.txt
flask
flask-cors
markdown
requests
google
python-dotenv